1. bug fixed: msconvert single file conversion to the same type => error status 2. add combined FDRsscore to FDR 3. add user-defined regular expression option to PIT:integrate

+++ trunk/wrappers/Proteomics/proteowizard/proteowizard.pl
@@ -11,13 +11,13 @@ my $path = $ARGV[$inputSize+4];
 my $os = lc($^O);
 print "OS: $os\n";
 #http://perldoc.perl.org/perlport.html#DOS-and-Derivatives
-my $removeCmd;
+my $removeCmd = "";
 if($os eq "mswin32" || $os eq "cygwin" || $os eq "dos" || $os eq "os2"){
 	my $tmpFolder = "C:\\lwr_staging\\tmp";
 	if($inputSize == 1){
 		system("copy $ARGV[0] $tmpFolder\\tmp_$id.$type");
 		system("msconvert $tmpFolder\\tmp_$id.$type --$outputType --outfile tmp_$id.$outputType -o $tmpFolder");
-		$removeCmd = "del $tmpFolder\\tmp_$id.$type";
+		$removeCmd = "del $tmpFolder\\tmp_$id.$type" unless ($type eq $outputType);
 	}else{
 		my $convertCmd = "msconvert";
 		$removeCmd = "del";
@@ -30,12 +30,12 @@ if($os eq "mswin32" || $os eq "cygwin" || $os eq "dos" || $os eq "os2"){
 		system($convertCmd);
 	}
 	system("move $tmpFolder\\tmp_$id.$outputType $output");
-	system($removeCmd);
+	system($removeCmd) if((length $removeCmd)!=0);
 }else{#linux
 	if($inputSize == 1){
 		system("cp $ARGV[0] /$path/tmp_$id.$type");
 		system("/$path/../../../gio_applications/proteowizard/msconvert /$path/tmp_$id.$type --$outputType --outfile tmp_$id.$outputType -o $path");
-		$removeCmd = "rm /$path/tmp_$id.$type";
+		$removeCmd = "rm /$path/tmp_$id.$type" unless ($type eq $outputType);
 	}else{
 		my $convertCmd = "/$path/../../../gio_applications/proteowizard/msconvert";
 		$removeCmd = "rm";
@@ -52,7 +52,7 @@ if($os eq "mswin32" || $os eq "cygwin" || $os eq "dos" || $os eq "os2"){
 	system("mv /$path/tmp_$id.$outputType $output");
 #	print("mv /$path/tmp_$id.$outputType $output\n");
 #	print "remove command: $removeCmd\n";
-	system($removeCmd);
+	system($removeCmd) if((length $removeCmd)!=0);
 }
 
 #print "Input file: $input\n";
